requirements.txt
aiofiles==0.7.0
aioredis==1.3.1
aiosmtplib==1.1.4
aiosqlite==0.16.1
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fakeredis==1.5.0
fastapi==0.63.0
fastapi-mail==0.3.5.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.12.3
httpx==0.17.1
idna==3.1
iso8601==0.1.14
Jinja2==2.11.3
MarkupSafe==1.1.1
passlib==1.7.4
Pillow==8.2.0
pkg-resources==0.0.0
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pypika-tortoise==0.1.0
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2020.5
redis==3.5.3
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.3.0
starlette==0.13.6
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
uvicorn==0.13.4